% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Shapes.R
\name{circleS3}
\alias{circleS3}
\title{Constructor for the circleS3 class}
\usage{
circleS3(radius, color = "blue")
}
\arguments{
\item{radius}{a numeric scalar denoting the radius of the circle}

\item{color}{a string denoting the color of the circle}
}
\value{
an object of class circleS3
}
\description{
Constructor for the circleS3 class
}
\examples{
myCircle <- circleS3(3, "yellow")
}
