% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Shapes.R
\name{print.circleS3}
\alias{print.circleS3}
\title{Print method for circleS3 class}
\usage{
\method{print}{circleS3}(x, ...)
}
\arguments{
\item{x}{An object of class circleS3}

\item{...}{ignored for this method}
}
\value{
A textual representation of the circleS3 object
}
\description{
Print method for circleS3 class
}
\examples{
print(circleS3(radius = 4))
}
