% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Shapes.R
\name{print.rectS3}
\alias{print.rectS3}
\title{Print method for rectS3 class}
\usage{
\method{print}{rectS3}(x, ...)
}
\arguments{
\item{x}{An object of class rectS3}

\item{...}{ignored for this method}
}
\value{
A textual representation of the rectS3 object
}
\description{
Print method for rectS3 class
}
\examples{
print(rectS3(width = 4, height = 3))
}
