% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Shapes.R
\name{print.shapeS3}
\alias{print.shapeS3}
\title{Print method for shapeS3 class}
\usage{
\method{print}{shapeS3}(x, ...)
}
\arguments{
\item{x}{An object of class shapeS3}

\item{...}{ignored for this method}
}
\value{
A textual representation of the shape object
}
\description{
Print method for shapeS3 class
}
\examples{
print(shapeS3("green"))
}
