% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Shapes.R
\name{print.triangleS3}
\alias{print.triangleS3}
\title{Print method for triangleS3 class}
\usage{
\method{print}{triangleS3}(x, ...)
}
\arguments{
\item{x}{An object of class triangleS3}

\item{...}{ignored for this method}
}
\value{
A textual representation of the triangleS3 object
}
\description{
Print method for triangleS3 class
}
\examples{
print(triangleS3())
}
