% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Shapes.R
\name{rectS3}
\alias{rectS3}
\title{Constructor for the rectS3 class}
\usage{
rectS3(width, height, color = "blue")
}
\arguments{
\item{width}{a numeric scalar denoting the width of the rectangle}

\item{height}{a numeric scalar denoting the height of the rectangle}

\item{color}{a string variable denoting the color of the rectangle}
}
\value{
an object of class rectS3
}
\description{
Constructor for the rectS3 class
}
\examples{
myRect <- rectS3(3, 4)
mySquare <- rectS3(3, 3)
}
