% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Shapes.R
\name{shapeS3}
\alias{shapeS3}
\title{Constructor for the shapeS3 class}
\usage{
shapeS3(color = "blue")
}
\arguments{
\item{color}{a string variable denoting the color of the shape}
}
\value{
an object of class rectS3
}
\description{
Constructor for the shapeS3 class
}
\examples{
myShapeS3 <- shapeS3("green")
}
