% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Shapes.R
\name{triangleS3}
\alias{triangleS3}
\title{Constructor for S3 triangle class}
\usage{
triangleS3(type = "right", color = "blue")
}
\arguments{
\item{type}{a string variable denoting the type of the triangle, right, isosceles, etc.}

\item{color}{a string variable denoting the color of the triangle}
}
\value{
an object of class triangleS3
}
\description{
Constructor for S3 triangle class
}
\examples{
myTriangle <- triangleS3("green", "isosceles")
}
